fix(conversation):#WB-3835 rename  component RecipientList
