+++ conversation/frontend/src/features/message/message-header/index.tsx
@@ -1,5 +1,5 @@
 import { Avatar, useDate, useDirectory } from '@edifice.io/react';
-import { MessageRecipientList } from '~/components/message-recipient-list';
+import { MessageRecipientList } from '~/components/MessageRecipientList/MessageRecipientList';
 import { useI18n } from '~/hooks';
 import { Message } from '~/models';
